storybook package.json versions, moving Checkbox, tailwind formatting
